Start tracking changes in doc
At the moment, indexes are parsed, though I don't know if we have it
hooked up to run. Mixins and their functions are being mixed into tables
and maybe subrecords. Subrecords get no special handling and we can't
talk to a database. But we're writing out scopes, and writing go files
and editing ast's.