New milestone: building the generated code
We can now generate fully valid Go code that will create scopes and
interfaces that apply to the interface of the models. Obviously we don't
deal with subrecord yet, and the code hasn't even been run once, but at
least it builds. Next step is rewriting the test_file.gp file to
exercise more of the doc featureset, then getting automated tests around
the example subpackage. Still plenty of things to do.