Add Count functions
Count just runs the count for distinct rows (using primary key column),
CountOf will run a count for the current column, you can add a
Distinct() call to run a distinct count of the current column. CountBy
will take a sql string and run the Count for it. The way it's
implemented is that Count calls PrimaryKeyColumn().Distinct().CountOf(),
and CountOf assembles the sql for the count, then calls CountBy() with
the string. That way we don't duplicate the sql code and get all three
tested by calling just Count.