implement and test pluck* (not pluck struct), prototypes also
PluckInt and PluckString are tested, also added a Pick function so you
can pluck a function value back. Also PluckTime was added, but it
isn't tested yet. PluckStruct wasn't worked on, I'm going to think on it
a bit more. Also UpdateSQL and Delete were written but haven't been
tested yet, waiting on SaveAll to be written so the tests are simpler to
write.