New fixes, mock out Scope updates, start AR-ish functions
Got the example tests running again, and write an early implementation
of struct saving, where you pass in the connection you want to use. That
was a record can be saved to multiple connections, for unknowable
reasons. Also write the Update and Set function prototypes for making
UPDATE statements from scopes.