Move example tests to forum, miscellaneous fixes
Parse had an example subdirectory that predated a lot of the code
organization that is present today (also parse is a bad name and needs
to change). I worked on a forum example for the upcoming join and
multi-table features, so it made sense to move the parse/example tests
into the forum directory. SaveAll was made as simple as possible, need
to support batching for new records at some point in the future maybe.