Subrecord time!
Create a subrecord example in forum, then start writing functionality
for it. Mappers are  in place, but you can't include them yet. Columns
to select are now pulled back from the mapper's, so that was a big step.
Including is figured out (for subrecords at least), but hasn't been
activated yet.