Compilation speed improvements for large schemas
This concludes my work on optimizing compilation speed as my 2500 line
schema is now building instantly. Lines dropped from 220k to 165k. Not
sure what exactly was causing the slow speeds, but the
simpleCols/simpleVals optimization made a huge improvement, in both
lines of code in the schema file, and compilation speed.