Great Interface Purge
After watching the dotGo Final Go Panel, I realized that I fell into one
of the Anti-Patterns for building these giant interfaces so I could hide
the scope structs. Since all the data is now in the internalScope
struct, I'm satisfied that the struct data is sufficiently cautioned.
This helped the generated code by quite a bit, lines dropped a bit, and
memory usage for large libraries dropped precipitously (5.5 to 1.7 on
2000+ column libraries).